finish controllers

+++ src/infrastructure/database/models/location_point.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, UniqueConstraint
+from sqlalchemy import Column, Integer, Float, UniqueConstraint, DECIMAL
 from src.infrastructure.database.models.base import Base
 
 
@@ -6,8 +6,8 @@ class LocationPointDB(Base):
     __tablename__ = 'location_points'
 
     id = Column(Integer, primary_key=True)
-    latitude = Column(Float, nullable=False)
-    longitude = Column(Float, nullable=False)
+    latitude = Column(DECIMAL, nullable=False)
+    longitude = Column(DECIMAL, nullable=False)
 
     __table_args__ = (
         UniqueConstraint('latitude', 'longitude', name='location_coordinates'),
